Made placeasker give more consistent results.
If part placing is considered good enough, reported color should be the
good one.
And conversely if placing is considered not good enough,
reported color should not be the good one. Chosen way to achieve this is
to compare the part displaced pixels with what is beneath. The
represented color(s) are collected with the count of pixels for each.
I introduced a second chance to consider the placing good enough
if sufficient a ratio of matching pixels is found.
If no such "good" ratio is found, we choose the most represented color
that is not the good answer.