placeasker fairer for parts not yet connected to any visible border.
If a part to place has no visible border yet in the current game, it is
more difficult to place than a part for which at least a portion of
border is "visible" (a neighbouring part with a common border is
already shown or the part has a border with a not clickable --and thus
already visible-- part).
One way to counterbalance this is to be more lenient on distance
criterium when the part to place has no visible neighbour than for a
place that has. So we have to know common borders. We get this by
scanning pixels at placeasker creation.