Add location name in results table
In "Location of Countries" quiz, the table of results displays only the
color of the user answer. As mentioned in bug report, in maps with many
countries, it could be useful to display the name of the location
selected by the user. So, results table has been enhanced displaying
location names as well.

BUGS: 141442