Implemented blurred flag feature
**What does this merge request address?**

This merge request relates to the Bug 457725, which suggested a new feature whereby certain flags with text on them can be blurred out, to make the flag guessing games more challenging. The flag should only be blurred in the guessing games, and in normal "browsing" mode the flag should remain unblurred. 

**How has the new functionality been implemented?**

The map parser has been amended to now look for an additional division property entitled "blurredflag", which should contain a blurred variant of the flag. A new method has been added to the "division" class entitled "getFlagForQuestion", which returns a blurred flag if one is available. This method is now used in the flag guessing games.

BUG: 457725